Rule for MissingInterpolator
I couldn't find a way in scala.reflect to get a symbol from a name, so the rule
is implemented as a ScopingRule.  Upon entering Templates or PackageDefs or
Blocks all the plausible declarations they carry are added to the scope.

This means that Blocks are processed as if their declarations were accessible from any point within the block.

Still need to add some code to ignore macros.